requirements.txt
Pillow==12.1.1
pyshorteners==1.0.1
selenium==4.41.0
